+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 ## In development
 
-- Add option to limit career choices by species career random table
 - Create a function to Prompt option to generate trappings or not
 - Generate actor with trapping
 - Create a function to Prompt a generation mode (simple or advanced)
@@ -13,6 +12,10 @@
   - A Object { [type: string]: {genDirectory: string, tokenIconPath: string}} to allow user create type of npc generation with specific path and icon
 - Create a function to Prompt select an npc type (used with settings to set a spcific generation path and a token icon path)
 - Create a simpliest generator with creature template
+
+## 1.1.0
+
+- Add a button "Random {Species}" on career chooser to limit random career allowed for selected species
 
 ## 1.0.0